Fix: PBS130 fix typo and change a sentence
Third paragraph, third sentence: "Codes doesn’t become well documented organically ...", "Codes" should be "Code".

In the "Longer is not a Synonym for Better!" section, 2nd paragraph: Changed "To abuse an old cliché goes ..." to "To abuse a Mark Twain witticism ,,,". (Otherwise, add the word "how" so to have "To abuse how an old cliché goes ...").